% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_for_good.R
\name{change_for_good}
\alias{change_for_good}
\title{Change a data frame, for good!}
\usage{
change_for_good(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{other arguments you'd probably pass to the mutate function}
}
\value{
A data frame with the modifications you specified
}
\description{
Create new variables in your data frame, changing it for good. (mutate equivalent)
}
\examples{
data <- read.csv("https://raw.githubusercontent.com/ashleymullan/wickedplyr/refs/heads/main/all_about_oz.csv")
data |> change_for_good(newfound_popularity = 2 * popularity)
}
