% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defy_gravity.R
\name{defy_gravity}
\alias{defy_gravity}
\title{Organize your rows so they defy gravity!}
\usage{
defy_gravity(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{other arguments you'd probably pass to the arrange function}
}
\value{
A data frame sorted on a specific variable
}
\description{
Pick a variable to sort by, and let your data frame defy gravity! (arrange equivalent)
}
\examples{
data <- read.csv("https://raw.githubusercontent.com/ashleymullan/wickedplyr/refs/heads/main/all_about_oz.csv")
data |> defy_gravity(swankifiedness) #
}
