% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toss_toss.R
\name{toss_toss}
\alias{toss_toss}
\title{Toss-toss the rows you don't need!}
\usage{
toss_toss(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{Logical predicates to keep rows}
}
\value{
A data frame with specified rows removed
}
\description{
Toss-toss the rows you don't need! (filter equivalent)
}
\examples{
data <- read.csv("https://raw.githubusercontent.com/ashleymullan/wickedplyr/refs/heads/main/all_about_oz.csv")
data |> toss_toss(accessory == "wand")
}
